+++ template-presets/theme/index.ts
@@ -1,0 +1,3 @@
+export { ThemeProvider } from "./provider";
+export { useTheme, useAppStyleSheet, AppStyleSheet } from "./provider";
+export { lightTheme, darkTheme } from "./themes";